Struct tcod_window::TcodWindow

pub struct TcodWindow {
    pub window: Rc<RefCell<Root>>,
    // some fields omitted
}

A window implemented by a TCOD back-end.

Fields

window

TCOD Root window used for rendering.

Methods

impl TcodWindow

fn new(settings: WindowSettings) -> Self

Create a new game window from the provided WindowSettings.

Internally, it initializes a new TCOD Root with the size and title specified in the WindowSettings.

Examples

use piston::window::{Size, WindowSettings};
use tcod_window::TcodWindow;

let mut window = TcodWindow::new(
    WindowSettings::new(
        "My Application".to_owned(),
        Size {
            width: 100,
            height: 100,
        }
    )
);

fn with_console(console: Rc<RefCell<Root>>, settings: WindowSettings) -> Self

Create a new game window from an existing TCOD Root console wrapped as an Rc<RefCell<Root>>.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use piston::window::{Size, WindowSettings};
use tcod::console::Root;
use tcod_window::TcodWindow;

let settings = WindowSettings::new(
    "My Application".to_owned(),
    Size {
        width: 100,
        height: 100,
    }
);
let root = Root::initializer()
                .size(settings.get_size().width as i32,
                      settings.get_size().height as i32)
                .title(settings.get_title())
                .init();
let console = Rc::new(RefCell::new(root));

let mut window = TcodWindow::with_console(console, settings);